+++ test/Extensions/Tester.Redis/Persistence/RedisPersistenceGrainTests.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
+using Orleans.Storage;
 using Orleans.TestingHost;
 using StackExchange.Redis;
 using Tester.Redis.Utility;
@@ -14,7 +15,7 @@ using Xunit.Abstractions;
 namespace Tester.Redis.Persistence
 {
     [TestCategory("Redis"), TestCategory("Persistence"), TestCategory("Functional")]
-    public partial class RedisPersistenceGrainTests : GrainPersistenceTestsRunner, IClassFixture<RedisPersistenceGrainTests.Fixture>
+    public class RedisPersistenceGrainTests : GrainPersistenceTestsRunner, IClassFixture<RedisPersistenceGrainTests.Fixture>
     {
         public static Guid ServiceId = Guid.NewGuid();
         public static string ConnectionStringKey = "ConnectionString";
@@ -42,10 +43,6 @@ namespace Tester.Redis.Persistence
                     hostBuilder.UseOrleans((ctx, siloBuilder) =>
                     {
                         siloBuilder
-                            //.UseRedisClustering(options =>
-                            //{
-                            //    options.ConnectionString = connectionString;
-                            //})
                             .AddRedisGrainStorage("GrainStorageForTest", options =>
                             {
                                 options.ConnectionString = connectionString;
@@ -75,6 +72,69 @@ namespace Tester.Redis.Persistence
                 StringValue = "string value",
                 GrainValue = fixture.GrainFactory.GetGrain<IGrainStorageGenericGrain<GrainState>>(999)
             };
+        }
+
+        // Redis specific tests
+
+        private GrainState state;
+        private IDatabase database;
+
+        [Fact]
+        public async Task Redis_InitializeWithNoStateTest()
+        {
+            var grain = fixture.GrainFactory.GetGrain<IGrainStorageGenericGrain<GrainState>>(0);
+            var result = await grain.DoRead();
+
+            //Assert.NotNull(result);
+            Assert.Equal(default(GrainState), result);
+            //Assert.Equal(default(string), result.StringValue);
+            //Assert.Equal(default(int), result.IntValue);
+            //Assert.Equal(default(DateTime), result.DateTimeValue);
+            //Assert.Equal(default(Guid), result.GuidValue);
+            //Assert.Equal(default(ITestGrain), result.GrainValue);
+        }
+
+        [Fact]
+        public async Task Redis_TestStaticIdentifierGrains()
+        {
+            var grain = fixture.GrainFactory.GetGrain<IGrainStorageGenericGrain<GrainState>>(12345);
+            await grain.DoWrite(state);
+
+            var grain2 = fixture.GrainFactory.GetGrain<IGrainStorageGenericGrain<GrainState>>(12345);
+            var result = await grain2.DoRead();
+            Assert.Equal(result.StringValue, state.StringValue);
+            Assert.Equal(result.IntValue, state.IntValue);
+            Assert.Equal(result.DateTimeValue, state.DateTimeValue);
+            Assert.Equal(result.GuidValue, state.GuidValue);
+            Assert.Equal(result.GrainValue, state.GrainValue);
+        }
+
+        [Fact]
+        public async Task Redis_TestRedisScriptCacheClearBeforeGrainWriteState()
+        {
+            var grain = fixture.GrainFactory.GetGrain<IGrainStorageGenericGrain<GrainState>>(1111);
+
+            await database.ExecuteAsync("SCRIPT", "FLUSH", "SYNC");
+            await grain.DoWrite(state);
+
+            var result = await grain.DoRead();
+            Assert.Equal(result.StringValue, state.StringValue);
+            Assert.Equal(result.IntValue, state.IntValue);
+            Assert.Equal(result.DateTimeValue, state.DateTimeValue);
+            Assert.Equal(result.GuidValue, state.GuidValue);
+            Assert.Equal(result.GrainValue, state.GrainValue);
+        }
+
+        [Fact]
+        public async Task Redis_DoubleActivationETagConflictSimulation()
+        {
+            var grain = fixture.GrainFactory.GetGrain<IGrainStorageGenericGrain<GrainState>>(54321);
+            var data = await grain.DoRead();
+
+            var key = grain.GetGrainId().ToString();
+            await database.HashSetAsync(key, new[] { new HashEntry("etag", "derp") });
+
+            await Assert.ThrowsAsync<InconsistentStateException>(() => grain.DoWrite(state));
         }
     }
 }
