+++ test/Extensions/Tester.Redis/Utility/TestExtensions.cs
@@ -19,16 +19,13 @@ namespace Tester.Redis.Utility
             var timeoutCancellationTokenSource = new CancellationTokenSource();
             var completedTask = await Task.WhenAny(taskToComplete, Task.Delay(timeout, timeoutCancellationTokenSource.Token));
 
-            // We got done before the timeout, or were able to complete before this code ran, return the result
             if (taskToComplete == completedTask)
             {
                 timeoutCancellationTokenSource.Cancel();
-                // Await this so as to propagate the exception correctly
                 await taskToComplete;
                 return;
             }
 
-            // We did not complete before the timeout, we fire and forget to ensure we observe any exceptions that may occur
             taskToComplete.Ignore();
             throw new TimeoutException(string.Format("WithTimeout has timed out after {0}.", timeout));
         }
